requirements.txt
# Python 项目依赖文件
# 安装命令: pip install -r requirements.txt
# 生成命令: pip freeze > requirements.txt

# ===== 集成框架 ====
requests==2.31.0
complexjson==1.2.0
PyMySQL==0.9.3
urllib3==2.4.0
PyYAML==6.0.3
ComplexJSON==1.2.0
iniconfig==2.3.0

# ===== 日志管理 =====
# loguru==0.7.2                 # 更友好的日志库
# structlog==23.1.0

# ===== 测试框架 =====
pytest==7.4.0
# pytest-asyncio==0.21.0
# pytest-cov==4.1.0
# pytest-html==4.0.2
# pytest-xdist==3.3.1           # 并行测试
# pytest-mock==3.11.1

# ==== 测试报告 ====
allure-pytest==2.15.2

# ===== 工具类 =====
# click==8.1.6                  # 命令行工具
# rich==13.5.2                  # 终端美化
# tqdm==4.65.0                  # 进度条
# python-dateutil==2.8.2        # 日期处理
# pytz==2023.3                  # 时区处理